Поправил замечания

+++ updater/src/main/java/ru/zubcov/paymentstats/updater/model/Payment.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.validation.constraints.PositiveOrZero;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -16,12 +17,12 @@ public class Payment {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @Column(name = "client_id", nullable = false)
-    private long clientId;
+    private Long clientId;
     @Column(name = "client_account", length = 20, nullable = false)
     private String clientAccount;
     @Column(name = "amount", nullable = false)
     @PositiveOrZero
-    private double amount;
+    private BigDecimal amount;
     @Column(name = "okved_category ", length = 20)
     private String okvedCategory;
     @Column(name = "receive_account", length = 20, nullable = false)
